Update remover.sh to handle sap_deployer removal
When checking for a valid az session, we missed redirecting the output to a file. It just throws an error to screen saying stdout.az not found.

The sap_deployer module does not implement the tf_state_resource_id parameter as the deployer state is migrated to azurerm backend, after initial provisioning. To prevent this from failing when removing the deployer we add an IF conditional to control the parameters passed to terraform destory.